fail if invalid ini file passed
When an non-existent .ini file is given on the command line, the builder
returns zero exit status with stdout:

 * OLPC OS builder v7.0.0
 * Build completed successfully.
 * Output is in /path/to/build/output

This was because RawConfigParser.read() is specified to ignore any files
that cannot be opened.

Follow the suggestion on ConfigParser documentation to call readfp
method on the first required file, and call read on subsequent files.

(Use case: subsequent .ini files are given to the builder for signing,
official, or build customisation option collections, in order to avoid
having to change and commit the master .ini file.)