sugar: explicitly depend on broadband providers
mobile-broadband-provider-info is needed to populate the lists of
countries and broadband service providers in the modem configuration
control panel.

The dependency is not declared by the current Sugar RPMs on OLPC
dropbox.

We can't rely on the Sugar RPMs, as the .spec files are apparently not
under any source control.  So add the dependency explicitly, marked as
probably not required from Fedora 21 and onwards.

See also
http://dev.laptop.org/ticket/12828
http://bugs.sugarlabs.org/ticket/4784
https://bugzilla.redhat.com/show_bug.cgi?id=1146534