base: lower ad-hoc default route preference
When an ad-hoc network is connected, and a USB ethernet adapter is
plugged in, the laptop cannot reach the internet.

Regression introduced by bdd624b.

Lower the preference for the ad-hoc default route to 2048, so that the
preference of 1024 used by the system for the USB ethernet adapter takes
precedence.

See #12817
http://dev.laptop.org/ticket/12817