sugar: add python-abiword
Write activity would fail to start (#12809), and a Python reproducer
demonstrated the same problem, with a segmentation fault #(abisource

This was caused by missing package python-abiword.

The dependency is not declared by the current Sugar RPMs on OLPC
dropbox.

We can't rely on the Sugar RPMs, as the .spec files are apparently not
under any source control.  It would seem inappropriate to use Sugar trac
to report it, since Sugar sources do not declare component dependencies.