base: stty icrnl on serial console
With the move to Fedora 20, shell and utility prompts on serial console
are not terminated by enter key, because the tty termios iflag ICRNL bit
is missing.

Might be caused by systemd, according to an upstream patch that is not
present in our build.  See util-linux patch
bb280f79f188f910c41e97fd5fa8889ae8b76d62.

Maintaining our own util-linux would be excessive for a low priority
problem.

Add a workaround in /root/.bashrc which is invoked only when serial
console is used, not when xterm or other shells start.

Fixes #12807.