zhashfs: improve performance on multi core systems
Do hash and compress in parallel threads, one for each core on the
system.  Use mutexes to protect input and output files from simultaneous
access.

Did increase compression performance by a factor of five on an eight
core x86_64 system.

However, as a result of the algorithm used, blocks in the output file
are no longer in increasing order.

The effect of this on an XO-4 using 41007o4.zd was a factor of three
degradation in fs-update time.

More work is needed to buffer the blocks before writing, to enforce
strict ordering.