yumcfg: allow new repositories
Adding repositories such as the OLPC restricted software or rpmfusion
results in a build where yum cannot be safely used, because packages are
not in olpc-exclude.

Adding these repositories to yumcfg does solve this, but rpmfusion is
using new repository format with hashed xml file names.

Use the new method of resolving file names in repositories.