sd_card_image: compile zhashfs
When zhashfs is not compiled, build continues for a long time before
reporting failure.

Better to fail early.

So if zhashfs is not compiled, try to compile it, and if that fails
then fail the build.

Not intended for use in the packaged form of olpc-os-builder, but since
most people are using olpc-os-builder from git this shouldn't matter.