xo1: reduce pulsing icon rate
The Sugar pulsing icon used during activity startup steals time from the
starting activity.  This has an effect on startup time.  The pulse rate
is ten times per second.

Reduce the rate of pulsing on XO-1 only, to twice a second.

In a shared laptop environment, this also signals to users the
performance that should be expected.

In a testing environment the number of pulses can be counted by
observation, and then halved to calculate the startup time.